:recycle:refactor: refactor posting detail page 중복 컴포넌트, mock data 통합 컴포넌트, 스타일 분리 Related to: #55
