+++ src/components/PostDetailPage/PostingContent.tsx
@@ -13,53 +13,56 @@ interface PostingContentProps {
   };
 }
 
+const styles = {
+  bookPreviewBox: {
+    width: '90%',
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center',
+    mt: '1rem',
+    padding: '0',
+  },
+  bookCard: {
+    width: '100%',
+    display: 'flex !important',
+    flexDirection: 'row !important',
+    backgroundColor: 'transparent',
+    padding: '1rem',
+    height: '60px',
+    '& .MuiCardMedia-root': {
+      width: '80px',
+      height: '100%',
+      padding: 'auto',
+      borderRadius: '0',
+    },
+    '& .MuiCardContent-root': {
+      flex: 1,
+      padding: '0.1rem',
+      display: 'flex',
+      flexDirection: 'column',
+      justifyContent: 'center',
+      boxSizing: 'border-box',
+    },
+    '& .MuiTypography-body1': {
+      fontSize: '14px',
+    },
+    '& .MuiTypography-body2': {
+      fontSize: '11px',
+    },
+  },
+};
+
 const PostingContent = ({ content, book }: PostingContentProps) => {
   return (
     <>
       {/* 책 정보 */}
-      <Box
-        sx={{
-          width: '90%',
-          display: 'flex',
-          justifyContent: 'center',
-          alignItems: 'center',
-          mt: '1rem',
-          padding: '0',
-        }}
-      >
+      <Box sx={styles.bookPreviewBox}>
         {book && (
           <CommonBookCard
             image={mockBooks[0].imageUrl}
             title={book.title}
             author={book.author}
-            sx={{
-              width: '100%',
-              display: 'flex !important',
-              flexDirection: 'row !important',
-              backgroundColor: 'transparent',
-              padding: '1rem',
-              height: '60px',
-              '& .MuiCardMedia-root': {
-                width: '80px',
-                height: '100%',
-                padding: 'auto',
-                borderRadius: '0',
-              },
-              '& .MuiCardContent-root': {
-                flex: 1,
-                padding: '0.1rem',
-                display: 'flex',
-                flexDirection: 'column',
-                justifyContent: 'center',
-                boxSizing: 'border-box',
-              },
-              '& .MuiTypography-body1': {
-                fontSize: '14px',
-              },
-              '& .MuiTypography-body2': {
-                fontSize: '11px',
-              },
-            }}
+            sx={styles.bookCard}
           />
         )}
       </Box>
