+++ src/components/PostDetailPage/PostingShareDialog/PostingShareDialog.tsx
