--- src/components/PostDetailPage/mockBookOtherPosting.ts
@@ -1,41 +1,0 @@
-export const mockBookOtherPosting = [
-  {
-    id: 1,
-    title: '금각사의 상징성에 대하여',
-    content:
-      '금각사가 상징하는 완벽한 아름다움과 주인공의 내면 세계가 충돌하는 방식이 인상적입니다. 특히 마지막 장면에서...',
-    user: {
-      id: 3,
-      name: '김문학',
-      avatarUrl: 'https://mui.com/static/images/avatar/1.jpg',
-    },
-    createdAt: '2024-01-19T09:00:00.000Z',
-    likeCount: 24,
-  },
-  {
-    id: 2,
-    title: '미시마 유키오의 섬세한 묘사',
-    content:
-      '불교 사찰을 배경으로 인간의 집착과 광기를 그려내는 작가의 필력이 돋보입니다. 특히 계절의 변화와 함께...',
-    user: {
-      id: 4,
-      name: '박독서',
-      avatarUrl: 'https://mui.com/static/images/avatar/2.jpg',
-    },
-    createdAt: '2024-01-17T09:00:00.000Z',
-    likeCount: 18,
-  },
-  {
-    id: 3,
-    title: '아름다움에 대한 집착의 비극',
-    content:
-      '주인공의 심리 변화가 섬세하게 묘사되어 있습니다. 완벽한 아름다움을 향한 집착이 어떻게 파멸로 이어지는지...',
-    user: {
-      id: 5,
-      name: '정서평',
-      avatarUrl: 'https://mui.com/static/images/avatar/3.jpg',
-    },
-    createdAt: '2024-01-15T09:00:00.000Z',
-    likeCount: 31,
-  },
-];
